+++ R/midfieldr.R
@@ -54,7 +54,15 @@ NULL
 # binding
 if (getRversion() >= "2.15.1") {
   utils::globalVariables(c(
-    ".", ":=", "..cols_we_want", ".SD",
-    "..columns", "..columns_we_want", "..keep_col"
+    ".", ":=", "..cols_we_want", ".SD", "..columns", "..columns_we_want",
+    "..keep_col",
+    #following are column names in the four midfielddata data sets
+    "id", "institution", "cip6", "race", "sex", "term", "term_degree",
+    "term_enter", "degree", "hours_transfer", "hours_term", "transfer",
+    "age", "us_citizen", "home_zip", "high_school", "sat_math", "sat_verbal",
+    "act_comp", "level", "standing", "coop", "hours_term_attempt", "gpa_term",
+    "hours_cumul_attempt", "hours_cumul", "gpa_cumul", "term_course",
+    "abbrev", "number", "section", "hours_course", "type", "pass_fail",
+    "grade", "faculty_rank"
   ))
 }